requirements-dev.txt
mock==2.0.0
PyHamcrest==1.9.0
pytest==3.2.5
